repo structure: rename internal/pkg/ to internal/
