+++ internal/provider/resource_repository.go
@@ -7,8 +7,8 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/docker/terraform-provider-docker/internal/pkg/hubclient"
-	"github.com/docker/terraform-provider-docker/internal/pkg/repositoryutils"
+	"github.com/docker/terraform-provider-docker/internal/hubclient"
+	"github.com/docker/terraform-provider-docker/internal/repositoryutils"
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
